requirements.txt
openai
flask
requests
python-dotenv
